refact: 태그 검색 후 친구 요청 api 분리 및 디자인 변경

+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/data/dto/friend/request/FriendsReqRequestDto.kt
@@ -1,5 +1,5 @@
 package com.droidblossom.archive.data.dto.friend.request
 
 data class FriendsReqRequestDto(
-    val friendIds : List<Int>
+    val friendIds : List<Long>
 )
